# docs/swagger.yaml
basePath: /api
consumes:
- application/json
definitions:
...
  Kind:
    description: Kind defines enum for kind of subscription
    type: string
    x-go-package: plexobject.com/formicary/internal/types
...
  Period:
    description: Period defines enum for period of subscription
    type: string
    x-go-package: plexobject.com/formicary/internal/types
  Policy:
    description: Policy defines enum for policy of subscription
    type: string
    x-go-package: plexobject.com/formicary/internal/types
...
  Subscription:
    description: Subscription defines subscription
    properties:
      active:
        description: Active flag
        type: boolean
        x-go-name: Active
      cpu_quota:
        description: CPUQuota  allowed cpu seconds
        format: int64
        type: integer
        x-go-name: CPUQuota
      created_at:
        description: CreatedAt creation time
        format: date-time
        type: string
        x-go-name: CreatedAt
      disk_quota:
        description: DiskQuota allowed disk Mbytes
        format: int64
        type: integer
        x-go-name: DiskQuota
      ended_at:
        description: EndedAt ended-at
        format: date-time
        type: string
        x-go-name: EndedAt
      id:
        description: |-
          gorm.Model
          ID defines UUID for primary key
        type: string
        x-go-name: ID
      kind:
        $ref: '#/definitions/Kind'
      organization_id:
        description: OrganizationID defines org
        type: string
        x-go-name: OrganizationID
      period:
        $ref: '#/definitions/Period'
      policy:
        $ref: '#/definitions/Policy'
      price:
        description: Price of subscription in cents
        format: int64
        type: integer
        x-go-name: Price
      remaining_cpu_quota:
        description: RemainingCPUQuota  cpu seconds
        format: int64
        type: integer
        x-go-name: RemainingCPUQuota
      remaining_disk_quota:
        description: RemainingDiskQuota disk Mbytes
        format: int64
        type: integer
        x-go-name: RemainingDiskQuota
      started_at:
        description: StartedAt started-at
        format: date-time
        type: string
        x-go-name: StartedAt
      updated_at:
        description: UpdatedAt update time
        format: date-time
        type: string
        x-go-name: UpdatedAt
      user_id:
        description: UserID defines user
        type: string
        x-go-name: UserID
    type: object
    x-go-package: plexobject.com/formicary/internal/types
  SystemConfig:
    properties:
      created_at:
        description: CreatedAt job creation time
        format: date-time
        type: string
        x-go-name: CreatedAt
      id:
        description: |-
          gorm.Model
          ID defines UUID for primary key
        type: string
        x-go-name: ID
      kind:
        description: Kind defines kind of config property
        type: string
        x-go-name: Kind
      name:
        description: Name defines name of config property
        type: string
        x-go-name: Name
      scope:
        description: Scope defines scope such as default or org-unit
        type: string
        x-go-name: Scope
      secret:
        description: Secret for encryption
        type: boolean
        x-go-name: Secret
      updated_at:
        description: UpdatedAt job update time
        format: date-time
        type: string
        x-go-name: UpdatedAt
      value:
        description: Value defines value of config property
        type: string
        x-go-name: Value
    title: SystemConfig defines internal configuration shared by the formicary platform.
    type: object
    x-go-package: plexobject.com/formicary/queen/types
...
host: https://formicary.io
info:
  contact:
    email: support@formicary.io
    name: Support
    url: https://formicary.io
  description: |-
    The formicary is a distributed orchestration engine based on `Leader-Follower`, `Pipes-Filter`, `Fork-Join` and `SEDA` design principles for
    executing a directed acyclic graph of tasks, which is also referred as a job workflow. A task represents a unit of work and a job definition is used to specify the task
    dependencies in the graph/workflow including configuration parameters and conditional logic.
  license:
    name: AGPL
    url: https://opensource.org/licenses/AGPL-3.0
  title: Formicary API.
  version: 0.0.1
paths:
...
  /api/configs:
    get:
      description: |-
        Queries system configs
        `This requires admin access`
      operationId: querySystemConfigs
      parameters:
      - format: int64
        in: query
        name: page
        type: integer
        x-go-name: Page
      - format: int64
        in: query
        name: page_size
        type: integer
        x-go-name: PageSize
      - description: Scope defines scope such as default or org-unit
        in: query
        name: scope
        type: string
        x-go-name: Scope
      - description: Kind defines kind of config property
        in: query
        name: kind
        type: string
        x-go-name: Kind
      - description: Name defines name of config property
        in: query
        name: name
        type: string
        x-go-name: Name
      responses:
        "200":
          $ref: '#/responses/sysConfigQueryResponse'
      tags:
      - system-configs
    post:
      description: '`This requires admin access`'
      operationId: postSystemConfig
      parameters:
      - in: body
        name: Body
        schema:
          $ref: '#/definitions/SystemConfig'
      responses:
        "200":
          $ref: '#/responses/sysConfigResponse'
      summary: Creates new system config based on request body.
      tags:
      - system-configs
  /api/configs/{id}:
    delete:
      description: '`This requires admin access`'
      operationId: deleteSystemConfig
      parameters:
      - in: path
        name: id
        required: true
        type: string
        x-go-name: ID
      responses:
        "200":
          $ref: '#/responses/emptyResponse'
      summary: Deletes an existing system config based on id.
      tags:
      - system-configs
    get:
      description: '`This requires admin access`'
      operationId: getSystemConfig
      parameters:
      - in: path
        name: id
        required: true
        type: string
        x-go-name: ID
      responses:
        "200":
          $ref: '#/responses/sysConfigResponse'
      summary: Finds an existing system config based on id.
      tags:
      - system-configs
    put:
      description: '`This requires admin access`'
      operationId: putSystemConfig
      parameters:
      - in: path
        name: id
        required: true
        type: string
        x-go-name: ID
      - in: body
        name: Body
        schema:
          $ref: '#/definitions/SystemConfig'
      responses:
        "200":
          $ref: '#/responses/sysConfigResponse'
      summary: Updates an existing system config based on request body.
      tags:
      - system-configs
...
  /api/subscriptions:
    get:
      description: '`This requires admin access`'
      operationId: querySubscriptions
      parameters:
      - format: int64
        in: query
        name: page
        type: integer
        x-go-name: Page
      - format: int64
        in: query
        name: page_size
        type: integer
        x-go-name: PageSize
      - description: Scope defines scope such as default or org-unit
        in: query
        name: scope
        type: string
        x-go-name: Scope
      - description: Kind defines kind of subscription property
        in: query
        name: kind
        type: string
        x-go-name: Kind
      - description: Name defines name of subscription property
        in: query
        name: name
        type: string
        x-go-name: Name
      responses:
        "200":
          $ref: '#/responses/subscriptionQueryResponse'
      summary: Queries system subscriptions.
      tags:
      - subscriptions
    post:
      description: '`This requires admin access`'
      operationId: postSubscription
      parameters:
      - in: body
        name: Body
        schema:
          $ref: '#/definitions/Subscription'
      responses:
        "200":
          $ref: '#/responses/subscriptionResponse'
      summary: Creates new system subscription based on request body.
      tags:
      - subscriptions
  /api/subscriptions/{id}:
    delete:
      description: '`This requires admin access`'
      operationId: deleteSubscription
      parameters:
      - in: path
        name: id
        required: true
        type: string
        x-go-name: ID
      responses:
        "200":
          $ref: '#/responses/emptyResponse'
      summary: Deletes an existing system subscription based on id.
      tags:
      - subscriptions
    get:
      description: '`This requires admin access`'
      operationId: getSubscription
      parameters:
      - in: path
        name: id
        required: true
        type: string
        x-go-name: ID
      responses:
        "200":
          $ref: '#/responses/subscriptionResponse'
      summary: Finds an existing system subscription based on id.
      tags:
      - subscriptions
    put:
      description: '`This requires admin access`'
      operationId: putSubscription
      parameters:
      - in: path
        name: id
        required: true
        type: string
        x-go-name: ID
      - in: body
        name: Body
        schema:
          $ref: '#/definitions/Subscription'
      responses:
        "200":
          $ref: '#/responses/subscriptionResponse'
      summary: Updates an existing system subscription based on request body.
      tags:
      - subscriptions
...
produces:
- application/json
responses:
...
  emptyResponse:
    description: Empty response body
...
  subscriptionQueryResponse:
    description: Query results of subscriptions
    schema:
      properties:
        Page:
          format: int64
          type: integer
        PageSize:
          format: int64
          type: integer
        Records:
          items:
            $ref: '#/definitions/Subscription'
          type: array
        TotalPages:
          format: int64
          type: integer
        TotalRecords:
          format: int64
          type: integer
      type: object
  subscriptionResponse:
    description: Subscription body for update
    schema:
      $ref: '#/definitions/Subscription'
  sysConfigQueryResponse:
    description: Query results of system-configs
    schema:
      properties:
        Page:
          format: int64
          type: integer
        PageSize:
          format: int64
          type: integer
        Records:
          items:
            $ref: '#/definitions/SystemConfig'
          type: array
        TotalPages:
          format: int64
          type: integer
        TotalRecords:
          format: int64
          type: integer
      type: object
  sysConfigResponse:
    description: SystemConfig body for update
    schema:
      $ref: '#/definitions/SystemConfig'
...
schemes:
- http
- https
security:
- api_key: []
securityDefinitions:
  api_key:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"
...